core: (Constraints) add AnyRange
